Rename Jetbrains Ignore-File and Add Some Symlinks, to Improve Visibility
This gitignore, only covers JetBrains products,
that are based on the IntelliJ Platform.
JetBrains also offers other products,
that are not base on IntelliJ (e. g. Resharper) and
therefore would need a totally different gitignore file.

So it seems to be more appropriate to name it according to
the platform instead of the company.

Also the names "JetBrains" or "IntelliJ" might not be obvious to somebody
looking for "GoLang" or "PhpStorm".
Adding some symlinks will hopefully solve the problem and
will make it easier for people to find what they are looking for.